adding azure function and removing  CCP
