+++ Solutions/AtlassianConfluenceAudit/Data Connector/AtlassianConfluenceAudit/ConfluenceAuditAPISentinelConnector/state_manager.py
