+++ Solutions/AtlassianConfluenceAudit/Data Connector/AtlassianConfluenceAudit/host.json
