+++ Solutions/AtlassianConfluenceAudit/Data/Solution_AtlassianConfluenceAudit.json
@@ -4,7 +4,7 @@
   "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Workbooks/Images/Logos/Azure_Sentinel.svg\"width=\"75px\"height=\"75px\">",
   "Description": "The [Atlassian Confluence Audit](https://www.atlassian.com/software/confluence) solution provides the capability to ingest [Confluence Audit Records](https://support.atlassian.com/confluence-cloud/docs/view-the-audit-log/) into Microsoft Sentinel.\n\n**Underlying Microsoft Technologies used:**\n\nThis solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs:\n\na. [Azure Monitor HTTP Data Collector API](https://docs.microsoft.com/azure/azure-monitor/logs/data-collector-api)\n\nb. [Azure Functions](https://azure.microsoft.com/services/functions/#overview)",
   "Data Connectors": [
-    "ConfluenceNativePollerConnector/azuredeploy_Confluence_native_poller_connector.json"
+    "Data Connector/AtlassianConfluenceAudit/ConfluenceAudit_API_FunctionApp.json"
   ],
   "Parsers": [
 	"Parsers/ConfluenceAudit.txt"
